requirements.txt
Flask==2.3.2
Werkzeug==2.3.7
Flask-JWT-Extended==4.5.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
python-dotenv==1.0.1
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
pytest==8.0.0
pytest-cov==4.1.0
redis==5.0.1
Sentry-SDK==1.40.6
uwsgi==2.0.23
